feat: dont use bind
