+++ src/middleware/node/types.ts
@@ -1,6 +1,8 @@
 import type { Logger } from "../../createLogger.js";
+import type { Webhooks } from "../../index.js";
 
 export type MiddlewareOptions = {
+  webhooks: Webhooks;
   path?: string;
   log?: Logger;
 };
